requirements.txt
gradio
numpy==1.23.5
torch==1.11.0
torchvision==0.12.0
openmim==0.1.5
mmdet==2.24.1
mmpose==0.25.1
